+++ CHANGES.md
@@ -3,6 +3,7 @@ New release
 
 - scrum board enhancements
 - minor bug fixes
+- add installation guide to README.md
 
 0.8.2 (13-05-2013)
 ===============